+++ idea/testData/codeInsight/generate/toString/common/dataClass.kt
@@ -1,2 +1,0 @@
-// NOT_APPLICABLE
-data class A<caret>(val n: Int)