+++ idea/testData/codeInsight/generate/toString/singleTemplate/dataClass.kt
@@ -1,0 +1,2 @@
+// GENERATOR: SINGLE_TEMPLATE
+data class A<caret>(val n: Int, val s: String?)